Add support for managing the agent install directory ACL on Windows
Prior to this revision, permissions for both *nix and Windows agents
were set to 0750 service_user:service_group. On Windows, this resulted
in an error when Puppet was running using an account other than
Administrator, even if the account was a member of the Administrators
group. This change explicitly grants full ACL access for $install_path
to Administrator, Administrators, and $service_user, and sets the owner
to Administrators. This allows both the Puppet install process and the
service to operate as expected.

- Resolves #2