Set ACL for $install_path after expanding installation archive
The archive module downloads the archive specified by $package_src and
sets ownership to the user running Puppet. When it extracts the archive,
it attempts to extract as $service_user, which fails due to the
restrictive ACL. Setting the ACL after the extraction fixes this issue
without creating any security issues, since the directory remains read
only (with no sensitive data) until the ACL is set.

In addition, the PowerShell dependency for Windows has been bumped to 5,
from 4. It was previously listed as 4 in error.